+++ self-api/src/main/java/com/selfxdsd/api/storage/Storage.java
@@ -1,13 +1,13 @@
 /**
  * Copyright (c) 2020, Self XDSD Contributors
  * All rights reserved.
- * <p>
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"),
  * to read the Software only. Permission is hereby NOT GRANTED to use, copy,
  * modify, merge, publish, distribute, sublicense, and/or sell copies of
  * the Software.
- * <p>
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -20,47 +20,35 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.selfxdsd.api;
+package com.selfxdsd.api.storage;
 
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.Test;
-import org.mockito.Mockito;
+import com.selfxdsd.api.ProjectManagers;
+import com.selfxdsd.api.Projects;
+import com.selfxdsd.api.Users;
 
 /**
- * Unit tests for {@link StoredProject}.
+ * Storage of Self.
  * @author Mihai Andronache (amihaiemil@gmail.com)
  * @version $Id$
  * @since 0.0.1
  */
-public final class StoredProjectTestCase {
+public interface Storage {
 
     /**
-     * StoredProject can return its Repo.
+     * Get the users Storage API.
+     * @return Users.
      */
-    @Test
-    public void returnsRepo() {
-        final Repo repo = Mockito.mock(Repo.class);
-        final Project project = new StoredProject(
-            repo, Mockito.mock(ProjectManager.class),
-            Mockito.mock(Storage.class)
-        );
-        MatcherAssert.assertThat(project.repo(), Matchers.is(repo));
-    }
+    Users users();
 
     /**
-     * StoredProject can return its ProjectManager.
+     * Get the project managers Storage API.
+     * @return ProjectManagers.
      */
-    @Test
-    public void returnsProjectManager() {
-        final ProjectManager manager = Mockito.mock(ProjectManager.class);
-        final Project project = new StoredProject(
-            Mockito.mock(Repo.class), manager,
-            Mockito.mock(Storage.class)
-        );
-        MatcherAssert.assertThat(
-            project.projectManager(),
-            Matchers.is(manager)
-        );
-    }
+    ProjectManagers projectManagers();
+
+    /**
+     * Get the projects Storage API.
+     * @return Projects.
+     */
+    Projects projects();
 }
