+++ self-api/src/main/java/com/selfxdsd/api/storage/StoredProjectManager.java
@@ -20,64 +20,74 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.selfxdsd.api;
+package com.selfxdsd.api.storage;
+
+import com.selfxdsd.api.Project;
+import com.selfxdsd.api.ProjectManager;
+import com.selfxdsd.api.Projects;
+import com.selfxdsd.api.Repo;
 
 /**
- * A Project stored in Self.<br><br>
+ * A Project Manager stored in Self. Use this class when implementing
+ * the Storage.<br><br>
  *
  * This class is in the API because it is implementation-agnostic!
  * It only works with API interfaces and nothing else.
  * @author Mihai Andronache (amihaiemil@gmail.com)
  * @version $Id$
  * @since 0.0.1
- * @todo #31:30min Implement the deactivate method which should remove the
- *  Project form the DB (it means Self will stop managing it). Return the
- *  corresponding Repo when done. Don't forget the tests.
  */
-public final class StoredProject implements Project {
+public final class StoredProjectManager implements ProjectManager {
 
     /**
-     * Repo of this Project.
+     * This PMs id.
      */
-    private final Repo repo;
+    private final int id;
 
     /**
-     * Manager in charge of this Project.
+     * This PM's access token.
      */
-    private final ProjectManager projectManager;
+    private final String accessToken;
 
     /**
-     * Self's Storage.
+     * Self's storage.
      */
     private final Storage storage;
 
     /**
      * Constructor.
-     * @param repo Repo of this project.
-     * @param projectManager Manager in charge.
-     * @param storage Storage of Self.
+     * @param id PM's id.
+     * @param accessToken API Access token.
+     * @param storage Self's storage.
      */
-    public StoredProject(
-        final Repo repo, final ProjectManager projectManager,
+    public StoredProjectManager(
+        final int id, final String accessToken,
         final Storage storage
     ) {
-        this.repo = repo;
-        this.projectManager = projectManager;
+        this.id = id;
+        this.accessToken = accessToken;
         this.storage = storage;
     }
 
     @Override
-    public ProjectManager projectManager() {
-        return this.projectManager;
+    public int id() {
+        return this.id;
     }
 
     @Override
-    public Repo repo() {
-        return this.repo;
+    public String accessToken() {
+        return this.accessToken;
     }
 
     @Override
-    public Repo deactivate() {
-        return null;
+    public Project assign(final Repo repo) {
+        final Project project = new StoredProject(repo, this, this.storage);
+        this.storage.projects().register(project);
+        return project;
+    }
+
+    @Override
+    public Projects projects() {
+        return this.storage.projects().assignedTo(this.id);
     }
 }
