+++ self-api/src/test/java/com/selfxdsd/api/storage/StoredProjectTestCase.java
@@ -1,13 +1,13 @@
 /**
  * Copyright (c) 2020, Self XDSD Contributors
  * All rights reserved.
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"),
  * to read the Software only. Permission is hereby NOT GRANTED to use, copy,
  * modify, merge, publish, distribute, sublicense, and/or sell copies of
  * the Software.
- *
+ * <p>
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -20,31 +20,50 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.selfxdsd.api;
+package com.selfxdsd.api.storage;
+
+import com.selfxdsd.api.Project;
+import com.selfxdsd.api.ProjectManager;
+import com.selfxdsd.api.Repo;
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.Test;
+import org.mockito.Mockito;
 
 /**
- * Storage of Self.
+ * Unit tests for {@link StoredProject}.
  * @author Mihai Andronache (amihaiemil@gmail.com)
  * @version $Id$
  * @since 0.0.1
  */
-public interface Storage {
+public final class StoredProjectTestCase {
 
     /**
-     * Get the users Storage API.
-     * @return Users.
+     * StoredProject can return its Repo.
      */
-    Users users();
+    @Test
+    public void returnsRepo() {
+        final Repo repo = Mockito.mock(Repo.class);
+        final Project project = new StoredProject(
+            repo, Mockito.mock(ProjectManager.class),
+            Mockito.mock(Storage.class)
+        );
+        MatcherAssert.assertThat(project.repo(), Matchers.is(repo));
+    }
 
     /**
-     * Get the project managers Storage API.
-     * @return ProjectManagers.
+     * StoredProject can return its ProjectManager.
      */
-    ProjectManagers projectManagers();
-
-    /**
-     * Get the projects Storage API.
-     * @return Projects.
-     */
-    Projects projects();
+    @Test
+    public void returnsProjectManager() {
+        final ProjectManager manager = Mockito.mock(ProjectManager.class);
+        final Project project = new StoredProject(
+            Mockito.mock(Repo.class), manager,
+            Mockito.mock(Storage.class)
+        );
+        MatcherAssert.assertThat(
+            project.projectManager(),
+            Matchers.is(manager)
+        );
+    }
 }
