+++ self-core-impl/src/main/java/com/selfxdsd/core/GithubRepo.java
@@ -23,7 +23,7 @@
 package com.selfxdsd.core;
 
 import com.selfxdsd.api.Project;
-import com.selfxdsd.api.Storage;
+import com.selfxdsd.api.storage.Storage;
 import com.selfxdsd.api.User;
 import java.net.URI;
