+++ self-core-impl/src/test/java/com/selfxdsd/core/GithubITCase.java
@@ -24,7 +24,7 @@ package com.selfxdsd.core;
 
 import com.selfxdsd.api.Project;
 import com.selfxdsd.api.Provider;
-import com.selfxdsd.api.Storage;
+import com.selfxdsd.api.storage.Storage;
 import com.selfxdsd.api.User;
 import com.selfxdsd.core.mock.InMemory;
 import org.hamcrest.MatcherAssert;
