+++ self-core-impl/src/test/java/com/selfxdsd/core/mock/InMemoryProjectManagers.java
@@ -23,6 +23,8 @@
 package com.selfxdsd.core.mock;
 
 import com.selfxdsd.api.*;
+import com.selfxdsd.api.storage.Storage;
+import com.selfxdsd.api.storage.StoredProjectManager;
 
 import java.util.HashMap;
 import java.util.Iterator;
